Fix closing client disrupting other's transaction
The following scenario caused a possible disruption in a transaction:
(all clients below are threads or processes talking to the same card
reader, in mode SCARD_SHARE_SHARED)

1. client A has a connection (SCardConnect) up, without transaction;
2. client B connects, then starts a transaction (SCardBeginTransaction);
3. client A terminates without calling SCardDisconnect;
4. client C connects and starts a transaction that is accepted by the
   pcscd daemon while client B is put on hold;
5. client C talks to the smart card;
6. when client C ends the transaction, B is allowed to continue, but its
   transaction has been disrupted.

At step 4, the transaction of client C should have been refused (or put
on hold), and client C should not have been let talking to the card
until client B ended its transacton.

The root defect is at step 3: during the cleanup of client A, the owner
of the current transaction is overwritten by zero.

This fix removes the overwriting of the owner of the current
transaction. Consequently, client C does not get its transaction until
client B calls SCardEndTransaction, and client B does not get
disrupted.